requirements.txt
pytest
pymysql
request
xlrd
selenium
docstring-parser
excel2json-3
any-case
psycopg2
mysql